ft_print2

+++ ft_putstr_octal.c
@@ -1,18 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_put_octal.c                                     :+:      :+:    :+:   */
+/*   ft_putstr_octal.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mbouanik <mbouanik@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/29 13:19:16 by mbouanik          #+#    #+#             */
-/*   Updated: 2017/03/29 13:29:27 by mbouanik         ###   ########.fr       */
+/*   Updated: 2017/04/10 19:14:16 by mbouanik         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void			ft_putstr_octal_assign(long long n)
+void			ft_putstr_octal_assign(unsigned int n)
 {
 	if (n > 0)
 		ft_putstr_octal_assign(n / 8);
@@ -20,7 +20,7 @@ void			ft_putstr_octal_assign(long long n)
 		g_str[g_p++] = n % 8 + '0';
 }
 
-void			ft_putstr_octal(long long n)
+void			ft_putstr_octal(unsigned int n)
 {
 	if (n == 0)
 	{
